requirements.txt
gunicorn==20.1.0
flask
numpy
Pillow
scikit-learn
tensorflow
tensorflow-cpu